introduce an option to negate the filtering on a queryset
i.e. make the filter not return events that match all given criteria,
but rather return events, that do not match the combination of given
criteria.

otherwise it is impossible to define a retrieve permission for a
combination of two criteria (i.e. when user "does not have permission"
to exclude events that match the given combination of criteria, e.g.
exclude events at site X and below magnitude Y)