fix magnitude slider, precision setting
precision has to be specified as number of digits, e.g. "1" for accuracy
to 1/10 of a magnitude step.

See
https://github.com/seiyria/bootstrap-slider/blob/master/README.md#options

"""
precision

float

number of digits after the decimal of step value

The number of digits shown after the decimal. Defaults to the number of
digits after the decimal of step value.
"""