correctly specify timezone of origin time in event modal
origin time will be displayed in time zone of browser, which will be
some local time in 99% of cases i guess, and no time zone specifier is
added by default in pre-defined localized date formats like "medium"
(btw. since it was not enclosed in quotes "medium" did not work
correctly till now anyway, time was not shown only the date). This adds
time zone specifier to the displayed string, so that it's not ambiguous
anymore. In angular >=1.3 it would be possible to specify that we want
UTC. see https://stackoverflow.com/a/29675847), but at least it's now
clear that it's displaying local time